requirements.txt
cachetools==4.1.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
dialogflow==1.0.0
Flask==1.1.1
google-api-core==1.17.0
google-auth==1.14.3
googleapis-common-protos==1.51.0
grpcio==1.29.0
gunicorn==20.0.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.13.0
twilio==6.33.1
urllib3==1.25.7
Werkzeug==0.16.0
